requirements-base.txt

torch
transformers
pdf2image
pillow
PyMuPDF
pdfplumber
